♻️ extracted download functions and unified code
